check if there is a process before trying to destroy it to avoid NPE

+++ eventstore-client/src/test/java/de/qyotta/eventstore/InMemoryEventstoreProvider.java
@@ -117,7 +117,9 @@ public class InMemoryEventstoreProvider {
    }
 
    public void stop() {
-      eventStoreProcess.destroyForcibly();
+      if (eventStoreProcess != null) {
+         eventStoreProcess.destroyForcibly();
+      }
    }
 
    public boolean isRunning() {
